Solution to Calculate Area and Perimeter of Rectangles

+++ Calculate Area and Perimeter of Rectangles@@67877d61dc107486ad5c274b/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 struct rectangle {
-    int l;
-    int b;
+    float l;
+    float b;
 };
 
 int main() {
